+++ src/main/java/io/github/poshjosh/ratelimiter/node/NodeImpl.java
@@ -30,13 +30,10 @@ final class NodeImpl<V> implements MutableNode<V> {
     
     private final Node<V> parent;
 
-    private final List<Node<V>> children;
-
     NodeImpl(String name, V value, Node<V> parent) {
         this.name = Objects.requireNonNull(name);
         this.value = value;
         this.parent = parent;
-        this.children = new LinkedList<>();
         if(parent != null) {
             if(this.equals(parent)) {
                 throw new IllegalArgumentException("A node may not be parent to itself");
@@ -50,23 +47,31 @@ final class NodeImpl<V> implements MutableNode<V> {
         }
     }
 
+    private List<Node<V>> _children = null;
     @Override
     public boolean addChild(Node<V> child) {
-        final Object ref = child.getParentOrDefault(null);
-        if(Objects.equals(ref, this)) {
-            if(!this.children.contains(child)) {
-                return this.children.add(child);
-            }else{
-                return false;
-            }
-        }else{
+        final Node<V> parentNode = child.getParentOrDefault(null);
+        if (!Objects.equals(parentNode, this)) {
             throw new UnsupportedOperationException();
         }
+        if (_children == null) {
+            _children = new ArrayList<>();
+            return _children.add(child);
+        }
+
+        if(!_children.contains(child)) {
+            return _children.add(child);
+        }
+
+        return false;
+    }
+    private List<Node<V>> children() {
+        return _children == null ? Collections.emptyList() : _children;
     }
 
     @Override
     public boolean anyMatch(Predicate<Node<V>> test) {
-        return test.test(this) || children.stream().anyMatch(child -> child.anyMatch(test));
+        return test.test(this) || children().stream().anyMatch(child -> child.anyMatch(test));
     }
 
     @Override
@@ -77,7 +82,7 @@ final class NodeImpl<V> implements MutableNode<V> {
     @Override
     public Node<V> copyTo(Node<V> parent) {
         final Node<V> newNode = Nodes.of(name, value, parent);
-        children.forEach(child -> child.copyTo(newNode));
+        children().forEach(child -> child.copyTo(newNode));
         return newNode;
     }
 
@@ -117,15 +122,15 @@ final class NodeImpl<V> implements MutableNode<V> {
     }
 
     public boolean hasChildren() {
-        return !children.isEmpty();
+        return !children().isEmpty();
     }
 
     @Override
-    public Node<V> getChild(int index) { return children.get(index); }
+    public Node<V> getChild(int index) { return children().get(index); }
 
     @Override
     public int getChildCount() {
-        return children.size();
+        return children().size();
     }
 
     /**
@@ -133,7 +138,7 @@ final class NodeImpl<V> implements MutableNode<V> {
      */
     @Override
     public List<Node<V>> getChildren() {
-        return Collections.unmodifiableList(children);
+        return Collections.unmodifiableList(children());
     }
 
     @Override
@@ -143,7 +148,7 @@ final class NodeImpl<V> implements MutableNode<V> {
         hash = 11 * hash + Objects.hashCode(this.value);
         hash = 11 * hash + Objects.hashCode(this.parent);
 // To avoid stackoverflow, use either parent or children, but not both.
-//        hash = 11 * hash + Objects.hashCode(this.children);
+//        hash = 11 * hash + Objects.hashCode(this.children());
         return hash;
     }
 
@@ -172,7 +177,7 @@ final class NodeImpl<V> implements MutableNode<V> {
             return false;
         }
 // To avoid stackoverflow, use either parent or children, but not both.
-//        if (!Objects.equals(this.children, other.children)) {
+//        if (!Objects.equals(this.children(), other.children())) {
 //            return false;
 //        }
         return true;